refactor: extracted calendar and related functions from index
