+++ edweiss-app/components/core/formatTime.tsx
@@ -1,0 +1,7 @@
+const formatTime = (minutes: number) => {
+    const hours = Math.floor(minutes / 60);
+    const mins = minutes % 60;
+    return `${hours}:${mins < 10 ? '0' : ''}${mins}`;
+};
+
+export default formatTime;
