+++ edweiss-app/components/core/getCurrentDay.tsx
@@ -1,0 +1,4 @@
+export const getCurrentDay = () => {
+    const now = new Date();
+    return now.getDay();
+};
