+++ edweiss-app/components/core/getCurrentTimeInMinutes.tsx
@@ -1,0 +1,5 @@
+export const getCurrentTimeInMinutes = () => {
+    const now = new Date();
+    return now.getHours() * 60 + now.getMinutes();
+};
+
